requirements.txt
bcrypt==4.3.0
greenlet==3.1.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
